API

API documentation

Index

ModelReduction.craig_bampton

ModelReduction.craig_bampton

ModelReduction.get_global_matrices

ModelReduction.guyan_reduction

ModelReduction.craig_bampton — Method.

craig_bampton(K, M, r, l, n)

Reduce the stiffness and mass matrices with the Craig-Bampton method. K = original stiffness matrix, M = original mass matrix, r = retained DOF:s, l = internal DOF:s, n = the number of modes to keep.

source

ModelReduction.craig_bampton — Method.

craig_bampton(K, M, c, r, l, n)

Reduce the stiffness, damping and mass matrices with the Craig-Bampton method. K = original stiffness matrix, M = original mass matrix, c = damping ratio, r = retained DOF:s, l = internal DOF:s, n = the number of modes to keep.

source

ModelReduction.get_global_matrices — Method.

get_global_matrices(cb)

Return global matrices K and M. If matrices are already defined to cb.properties.K and cb.properties.M, return those instead.

source

ModelReduction.guyan_reduction — Method.

guyan_reduction(K, m, s)

Reduce the stiffness matrix by Guyan Reduction. K = original stiffness matrix, m = master nodes, s= slave nodes.

source