API

API documentation

Index

ModelReduction.craig_bampton

ModelReduction.global_mass

ModelReduction.global_stiffness

ModelReduction.guyan_reduction

ModelReduction.sort_nodes

ModelReduction.craig_bampton — Method.

craig_bampton(K, M, r, l, n)

Reduce the stiffness and mass matrices with the Craig-Bampton method. K = original stiffness matrix, M = original mass matrix, r = retained DOF:s, l = internal DOF:s, n = the number of modes to keep.

source

ModelReduction.global_mass — Method.

global_mass(m, N)

Calculate global mass matrix of size N.

source

ModelReduction.global_stiffness — Method.

global_stiffness(k, N)

Calculate global stiffness matrix of size N.

source

ModelReduction.guyan_reduction — Method.

guyan_reduction(K, m, s)

Reduce the stiffness matrix by Guyan Reduction. K = original stiffness matrix, m = master nodes, s= slave nodes.

source

ModelReduction.sort_nodes — Method.

sort_nodes(nodes, node_sets)

Create the r and l arrays for the retained and internal nodes. nodes = all nodes of the model node_sets = node sets containing the nodes that are to be retained

source